Adding ad sequence and pod sequence fields
As part of the OpenRTB 2.6 effort, specifying value lists for the 'sequence' and 'podsequence' fields, used with "audio" and "video" objects. These should be added to the VideoPlacement and AudioPlacement objects in AdCOM 1.0, as well. I assume that will be done in a separate PR once these lists are approved for use in OpenRTB 2.6 and AdCOM 1.0.